requirements.txt
AutoHotkey.py==0.1.1
numpy==1.21.2
Pillow==8.3.2
